Sad path image

+++ spec/requests/api/v1/forcast_request_spec.rb
@@ -41,7 +41,7 @@ describe "Forecast API endpoint" do
     expect(current_weather).to have_key(:feels_like)
     expect(current_weather[:feels_like]).to be_a(Float)
     expect(current_weather).to have_key(:humidity)
-    expect(current_weather[:humidity]).to be_an(Integer)
+    expect(current_weather[:humidity]).to be_an(Numeric)
     expect(current_weather).to have_key(:uvi)
     expect(current_weather[:uvi]).to be_a(Numeric)
     expect(current_weather).to have_key(:visibility)
